pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-quality-compliance-check"
version = "0.1.0"
description = "Document Quality & Compliance Check System"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.9",
    "reportlab>=4.1.0",
    "python-docx>=1.1.0",
    "pypdf>=5.0.0",
    "structlog>=24.1.0",
    "anthropic>=0.18.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "jinja2>=3.1.3",
    "bleach>=6.1.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "PyYAML>=6.0",
    "aamad",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true

